Preserve empty PATH_INFO for mounts
Allow PATH_INFO to remain an empty string when a request is made for the exact prefix name of a mount, e.g. a request to "/settings" for this route:

```
mount ->(env) { [200, {}, [env["PATH_INFO"]] }, at: "/settings"
```

This ensures that the Rack environment can be used to reconstruct the same path as used for the original request (e.g. "/settings" instead of "/settings/" for the example above).

To ensure compatibility for routers used within mounts, consider an empty string to be equivalent to "/" for the purposes of route matching.

This allows a router to be mounted and have it root route be used when a request is made to the exact mount prefix without any trailing slash:

```
mounted_router = Hanami::Router.new do
  root to: ->(env) { [200, {}, ["Hello from root"]] }
end

mount mounted_router, at: "/settings"
```

In the above example, a request to /settings will still serve the route returning "Hello from root".